include check for valid firmware

+++ update/atmega_firmware_upload.sh
@@ -15,48 +15,52 @@ if [ ! -d $openenergymonitor_dir/data/firmware ]; then
 fi
 
 result=$(./get_firmware_download_url.py $firmware_key)
-result=($result)
+if [ "$result" != "firmware not found" ]; then
+  result=($result)
 
-download_url=${result[0]}
-baud_rate=${result[1]}
+  download_url=${result[0]}
+  baud_rate=${result[1]}
 
-hexfile=$openenergymonitor_dir/data/firmware/$firmware_key.hex
+  hexfile=$openenergymonitor_dir/data/firmware/$firmware_key.hex
 
-echo "Downloading firmware from: "
-echo $download_url
-wget -q $download_url -O $hexfile
+  echo "Downloading firmware from: "
+  echo $download_url
+  wget -q $download_url -O $hexfile
 
-echo
-echo "Downloaded file: "
-ls -lh $hexfile
+  echo
+  echo "Downloaded file: "
+  ls -lh $hexfile
 
-if [ -f $hexfile ]; then
+  if [ -f $hexfile ]; then
 
-  state=$(systemctl show emonhub | grep ActiveState)
-  
-  if [ $state == "ActiveState=active" ]; then
+    state=$(systemctl show emonhub | grep ActiveState)
+    
+    if [ $state == "ActiveState=active" ]; then
+      echo
+      echo "EmonHub is running, stopping EmonHub"
+      sudo systemctl stop emonhub
+      emonhub_stopped_by_script=1
+    else
+      emonhub_stopped_by_script=0
+    fi
+    
     echo
-    echo "EmonHub is running, stopping EmonHub"
-    sudo systemctl stop emonhub
-    emonhub_stopped_by_script=1
+    echo "Uploading $firmware_key on serial port $serial_port"
+    echo
+    avrdude -v -c arduino -p ATMEGA328P -P /dev/$serial_port -b $baud_rate -U flash:w:$hexfile
+
+    echo
+    echo "Upload complete"
+    
+    if [ $emonhub_stopped_by_script == 1 ]; then
+      echo
+      echo "Restarting EmonHub"
+      sudo systemctl start emonhub
+    fi
+
   else
-    emonhub_stopped_by_script=0
+    echo "Firmware download failed...check network connection"
   fi
-  
-  echo
-  echo "Uploading $firmware_key on serial port $serial_port"
-  echo
-  avrdude -v -c arduino -p ATMEGA328P -P /dev/$serial_port -b $baud_rate -U flash:w:$hexfile
-
-  echo
-  echo "Upload complete"
-  
-  if [ $emonhub_stopped_by_script == 1 ]; then
-    echo
-    echo "Restarting EmonHub"
-    sudo systemctl start emonhub
-  fi
-
 else
-  echo "Firmware download failed...check network connection"
+  echo "Firmware not found: $firmware_key"
 fi
